V2 MCU optomize

+++ MCU_code/V2/Battery_Charging/Batter_Voltage_Check.ino
@@ -4,7 +4,7 @@ void BVC(double BV){
   Serial.print("Battery Voltage:");
   Serial.print(BV);
   Serial.print(" V\t");
-  int per = ((BV-11.4)*100/2.1+0.5);
+  short per = ((BV-11.4)*100/2.1+0.5);
   Serial.print((BV-11.4)*100/2.1);
   if (per==100) screen.setCursor(3,48);
   else if(per>=10) screen.setCursor(17,48);
