+++ MCU_code/V2/Battery_Charging/buzzer.ino
@@ -1,5 +1,5 @@
-void buzzer(int frequency,int buzzetime){
-  for(int i=0; i<buzzetime;i++){
+void buzzer(short frequency,short buzzetime){
+  for(short i=0; i<buzzetime;i++){
     analogWrite(6,frequency);
     delay(100);
     analogWrite(6,0);
